requirements.txt
# Web Framework & Server
fastapi==0.116.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Data Handling & ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
sentence-transformers==3.0.1
openpyxl==3.1.5
pyarrow==16.1.0


# Authentication & Security
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
google-auth==2.31.0
requests==2.32.3
email-validator==2.2.0

# Utilities
pydantic==2.8.2
python-dotenv==1.0.1
prisma==0.15.0
openai==1.35.10

